jest: fail on error messages
Use the library https://github.com/ValentinH/jest-fail-on-console to
spy on the errors to make test failing if any occur.

Transfer the configuration from the previous FilteredConsole to this new
library.

Allow a few targeted errors to occur to give some time to refactor while
keeping the benefits to have failing on new error messages.