router: add zero state to image builder
This commit introduces the zero state to the image builder application.
When the user doesn't have any composes in the image-builder data base,
they are navigated to a welcome page.

This welcome page exposes a button to create an image.
When the user gets to click on the button, the wizard opens up in a
modal and the user can create their first image.

Once the creation is under going, the user is navigated back to the
images table where they can track progress of the image being built up.

For more details: https://miro.com/app/board/uXjVM7duAEY=/?share_link_id=601171373473
Fixes HMS-2184