Revert "API: Add Compliance API slice"
This reverts commit 36f9e7056543f2e780b46b4d8e100ee33324bb7e.

The Compliance v1 API is not stable and not intended for consumption by
other teams. We will implement the OpenSCAP MVP without the Compliance
API for now with a hard coded list of polices in image-builder, and when
their v2 API is ready in the near future begin using it to retrieve
policies.