fsc: fix the oscap fs customizations
There was a bug preventing all the partitions required by an oscap
policy to show up in the fsc step. That bug was because the FSC
customizations were only accepted if they matched exactly the fix set of
base partitions directories.

So now there's a bit of string manipulation to avoid that but it's all
patched up.

fixes https://github.com/RedHatInsights/image-builder-frontend/issues/1385