devDeps: Remove eslint-config-react-app dependency
A dependabot PR https://github.com/RedHatInsights/image-builder-frontend/pull/1394 was failing with the following message:
"ESLint couldn't determine the plugin "testing-library" uniquely."

This was caused by having two separate instances of "eslint-plugin-testing-library" in dependencies. One directly in the `devDependencies` and another one as a dependency of eslint-config-react-app.

Seeing as [eslint-config-react-app](https://www.npmjs.com/package/eslint-config-react-app) was not updated in the last two years, the majors versions of deps in the "bundle" were crashing with the `devDependencies` one.