Blueprints: add blueprint feature
Blueprints can now be created through the Create Blueprint wizard. This is
almost identical to the Create Image wizard except for a couple blueprint specific
fields.

Created blueprints are shown in the blueprints sidebar where they can be filtered
or selected.

This commit also includes improvements to the general UI per the recent design
updates. The header is now Images, filtering has been added to the images table,
and the Create Image button is moved to the header.

Given that in image-builder blueprints are still experimental these new features
can be seen when running stage-beta:msw+experimental. The fixtures mimic blueprints
even though the cannot currently be fetched from the service.