wizard/common: introduce TypeAheadSelect
Based on the patternfly pattern pattern
https://www.patternfly.org/components/menus/select#typeahead, can be
used by components to spawn select list that can get filtered by the
user input. The component is also navigable by keyboard strokes.

It requires in input a list of elements to be displayed as select
options and a state+updateFunction to communicate the selected value.