wizard/common: introduce ValidatedTextField
Based on the patternfly pattern https://www.patternfly.org/components/forms/text-input
this component acts as a text field with extra validation. When the user
types in the answer, the input gets validates against a validating
function.

The component takes in as parameters a state+updateFunction to
communicate the entered value and also a validating function to check if
the entered value is up to specs.