wizard/targetEnvironment: add the AWS Target step
This commit adds the AWS target substep in the TargetEnvironment step.

It also introduces the SourcesSelect component. This component is based
on the patternfly pattern
https://www.patternfly.org/components/menus/select#typeahead

As opposed to the TypeAheadSelect, this component fetches the data to
populate the list itself. This is required because of the behavior of
the wizard: when the user clicks on the drop down menu, the list of
elements it contains gets update.