wizard/context: moving from props to context
The team made a call to use a context to hold the state of the image
wizard. This commits moves all the props that were passed as function
arguments and stores them in the context state.

The main advantages of such a design change is a less cluttered code
base as props don't have to be explicitly passed down. The main
inconvenient is that future changes won't be able to rely as much on the
language to assert that no step is changing the state of another one.
The team will have to monitor more closely new state additions.

The following changes were also made in this commit:
- No isFecthing or isError is being stored in the global context. If a
  component needs access to these information it needs to do the data
  call itself.
- Splitting of the CreateImageWizard.tsx file into 3 files,
  CreateImageWizard for the wizard, ImageWizardContext for the context
  and CustomWizardFooter for the custom footer.