V2Wizard: Add listeners for `<TargetEnvironment>` in Image Output step
Listeners are used to monitor changes to the release (aka distribution)
and architecture. If either of these values is changed, the listener
will correct the list of target environments by removing unsupported
targets. e.g. aarch64 images do not support azure, if azure is a
selected target and the arch is changed from x86 to aarch64, azure will
be removed from the list of selected targets.