LandingPage: Hotfix to remove exposed blueprint buttons
Commit bae6435fd92797e1879ff6f2798c11819e1dbff0 adds scaffolding for the
blueprint table. Blueprint features should be hidden behind experimental
flags. The new buttons related to creating blueprints and rebuilding
images in the header were not hidden behind the experimental flag, this
commit fixes that and does so.

bae6435fd92797e1879ff6f2798c11819e1dbff0 also changed the header title
from "Image Builder" to "Images". This change has been a long time
coming - it is being left in for now. We may want to revisit the copy in
the popovers to align it as the term "Image Builder" is still being used
in those.