wizard: disable the unbuildable oscap profiles
Some Oscap profiles can't get built. That is because they require to
disable a set of services that aren't installed on the system at this
moment.

A workaround is being worked on the backend side. But meanwhile, we
decide to disable this set of profiles to insure the user doesn't
stumble upon bound-to-fail configurations.