devDeps: Bump @testing-library/dom from 9.3.4 to 10.0.0
This bump was blocked by eslint-plugin-jest-dom dependency. I've noticed we don't use this set of rules anywhere and we don't really have a problem with breaking them in our code base. So the dependency was removed in favor of bumping @testing-library/dom.

Should it be needed in the future, it can be easily installed back and added to our eslint configs.