Wizard: add partitioning modes into file system config
The image builder API supports partitioning modes for a while.
Currently, the frontend doesn't specify one, which makes the backend
default to "auto-lvm". This partitions the image as lvm if the request
contains an extra mountpoint that isn't specified in the base partition
table. Otherwise, it's raw-partitioned. In other words, it's confusing.

This commit gives users an option to specify the mode explicitly: either
to lvm, or raw. This should reduce confusion and give users more power.

The partitioning mode is now enforced by the frontend. This means that
old blueprints are "converted" when edited: A simple heurestic is in
charge of that: if there's a custom mountpoint in the original
blueprint, the mode is set to lvm. Otherwise, it defaults to raw.