src: move the first boot service into constants.ts
I don't think there's a value in having the unit encoded in base64.
It looks fishy, and it's error prone. We discovered that the
custom-first-boot.service file in aux was different than the base64-
encoded string. There was nothing malicious, but we apparently forgot
to rerun the Makefile target.

Instead, we can just put the raw script into constants.js and base64-
encode it on the fly. The cost is close to zero.

This way, we can also get rid of the aux dir, and of the Makefile
target. 🧹