Konflux: Add ci.sh with workaround for Konflux pipeline
The Konflux pipeline clones the repo twice, making a nested clone
under ./source/source which causes webpack to break when the Cockpit
webpack configuration is present. `source ci.sh` is run at the
beginning of the Konflux pipeline and as a workaround for now we
simply remove everything related to Cockpit - this is fine as none of
it is needed for the service to run.