cockpit: add cockpit dependencies
The on-prem version of this repo depends on the `Cockpit` project. Some
of the files from the `Cockpit` repo are needed to build and run this
project. These files are saved into the `pkg/lib` directory, so we need
ot add some plumbing for this into the project.

Additionally, this commit adds some convenience scripts to the Makefile.