ci.sh: inject debug ids using sentry-cli
Since we're building with konflux now, we lost the debug ids. By
injecting the debug ids in konflux we regain unminified sourcemaps in
sentry. Note that jenkins will continue to upload the sourcemaps, which
are tied to specific releases, we just need the debug ids to be in the
production artefacts.